move signature_v4.go to pkg/s3
Have the code in a central place to make it accessible to additional
callers.

Also get the code for getContentSha256Cksum() from
service/proxy/auth/middleware.go as it is only called by
ComputeSignatureV4() and nowhere else.

Recommended git option to view this commit: --color-moved.
